// LifeLine/LifeLine.UI/src/app/modules/authentification/models/register-dto.ts
export interface RegisterDto {
    fullName: string;
    securityNumber: string;
    country: string;
    city: string;
    gender: string;
    birthDay: string;
    bloodType: number;
    rhFactor: number;
    phoneNumber: string;
    email: string;
    userName: string;
    password: string;
    confirmPassword: string;
}